Move goToSleep and button-only sleep into sleep_session
Relocate deep-sleep entry (timer + GPIO wake, X IQS prepare/wake stub,
config_gpio_for_lp) out of bl.cpp. Move startup_time into globals for
awake-time stats. Export submitStoredLogs for the sleep path.

wifiErrorDeepSleep stays in bl for a follow-up. Behavior unchanged.

Verified: pio run -e trmnl
